Fixed vex2difx problem where run failing because a job had no media.
In this instance the media changed between scans at KP.  The logic
created a split job of about 2 milliseconds duration.  This splinter job
occured between the end of the previous scan and the media change and
thus there was media at KP for those 2 ms and hence the error.

Modified the code in JobGroup::createJobs to check for splinter jobs.
The existing logic checked that the duration was positive and I
replaced it with a test against the minimum job duration (defaults to
2 seconds).

It's possible that this is a bandaid fix, but I think the problem
arises because of rounding error in the time.  I'm surprised it's
never been reported before.